+++ src/include/smash/listmodus.h
@@ -214,10 +214,10 @@ class ListModus : public ModusDefault {
   /// file_id_ is the id of the current file
   int file_id_;
 
-  /// Counter for mass-check warnings to avoid spamming
-  int n_warns_precision_ = 0;
-  /// Counter for energy-momentum conservation warnings to avoid spamming
-  int n_warns_mass_consistency_ = 0;
+  /// Auxiliary flag to warn about mass-discrepancies only once per instance
+  bool warn_about_mass_discrepancy_ = true;
+  /// Auxiliary flag to warn about off-shell particles only once per instance
+  bool warn_about_off_shell_particles_ = true;
 
   /// last read position in current file
   std::streampos last_read_position_;